+++ src/PhpSpreadsheet/Style/Style.php
@@ -120,10 +120,8 @@ class Style extends Supervisor
     /**
      * Get the shared style component for the currently active cell in currently active sheet.
      * Only used for style supervisor.
-     *
-     * @return Style
      */
-    public function getSharedComponent()
+    public function getSharedComponent(): self
     {
         $activeSheet = $this->getActiveSheet();
         $selectedCell = $this->getActiveCell(); // e.g. 'A1'
@@ -199,12 +197,12 @@ class Style extends Supervisor
      * );
      * </code>
      *
-     * @param array $pStyles Array containing style information
-     * @param bool $pAdvanced advanced mode for setting borders
+     * @param array $styleArray Array containing style information
+     * @param bool $advancedBorders advanced mode for setting borders
      *
      * @return $this
      */
-    public function applyFromArray(array $pStyles, $pAdvanced = true)
+    public function applyFromArray(array $styleArray, $advancedBorders = true)
     {
         if ($this->isSupervisor) {
             $pRange = $this->getSelectedCells();
@@ -237,36 +235,36 @@ class Style extends Supervisor
             }
 
             // ADVANCED MODE:
-            if ($pAdvanced && isset($pStyles['borders'])) {
+            if ($advancedBorders && isset($styleArray['borders'])) {
                 // 'allBorders' is a shorthand property for 'outline' and 'inside' and
                 //        it applies to components that have not been set explicitly
-                if (isset($pStyles['borders']['allBorders'])) {
+                if (isset($styleArray['borders']['allBorders'])) {
                     foreach (['outline', 'inside'] as $component) {
-                        if (!isset($pStyles['borders'][$component])) {
-                            $pStyles['borders'][$component] = $pStyles['borders']['allBorders'];
+                        if (!isset($styleArray['borders'][$component])) {
+                            $styleArray['borders'][$component] = $styleArray['borders']['allBorders'];
                         }
                     }
-                    unset($pStyles['borders']['allBorders']); // not needed any more
+                    unset($styleArray['borders']['allBorders']); // not needed any more
                 }
                 // 'outline' is a shorthand property for 'top', 'right', 'bottom', 'left'
                 //        it applies to components that have not been set explicitly
-                if (isset($pStyles['borders']['outline'])) {
+                if (isset($styleArray['borders']['outline'])) {
                     foreach (['top', 'right', 'bottom', 'left'] as $component) {
-                        if (!isset($pStyles['borders'][$component])) {
-                            $pStyles['borders'][$component] = $pStyles['borders']['outline'];
+                        if (!isset($styleArray['borders'][$component])) {
+                            $styleArray['borders'][$component] = $styleArray['borders']['outline'];
                         }
                     }
-                    unset($pStyles['borders']['outline']); // not needed any more
+                    unset($styleArray['borders']['outline']); // not needed any more
                 }
                 // 'inside' is a shorthand property for 'vertical' and 'horizontal'
                 //        it applies to components that have not been set explicitly
-                if (isset($pStyles['borders']['inside'])) {
+                if (isset($styleArray['borders']['inside'])) {
                     foreach (['vertical', 'horizontal'] as $component) {
-                        if (!isset($pStyles['borders'][$component])) {
-                            $pStyles['borders'][$component] = $pStyles['borders']['inside'];
+                        if (!isset($styleArray['borders'][$component])) {
+                            $styleArray['borders'][$component] = $styleArray['borders']['inside'];
                         }
                     }
-                    unset($pStyles['borders']['inside']); // not needed any more
+                    unset($styleArray['borders']['inside']); // not needed any more
                 }
                 // width and height characteristics of selection, 1, 2, or 3 (for 3 or more)
                 $xMax = min($rangeEndIndexes[0] - $rangeStartIndexes[0] + 1, 3);
@@ -315,7 +313,7 @@ class Style extends Supervisor
                         $range = $colStart . $rowStart . ':' . $colEnd . $rowEnd;
 
                         // retrieve relevant style array for region
-                        $regionStyles = $pStyles;
+                        $regionStyles = $styleArray;
                         unset($regionStyles['borders']['inside']);
 
                         // what are the inner edges of the region when looking at the selection
@@ -327,8 +325,8 @@ class Style extends Supervisor
                                 case 'top':
                                 case 'bottom':
                                     // should pick up 'horizontal' border property if set
-                                    if (isset($pStyles['borders']['horizontal'])) {
-                                        $regionStyles['borders'][$innerEdge] = $pStyles['borders']['horizontal'];
+                                    if (isset($styleArray['borders']['horizontal'])) {
+                                        $regionStyles['borders'][$innerEdge] = $styleArray['borders']['horizontal'];
                                     } else {
                                         unset($regionStyles['borders'][$innerEdge]);
                                     }
@@ -337,8 +335,8 @@ class Style extends Supervisor
                                 case 'left':
                                 case 'right':
                                     // should pick up 'vertical' border property if set
-                                    if (isset($pStyles['borders']['vertical'])) {
-                                        $regionStyles['borders'][$innerEdge] = $pStyles['borders']['vertical'];
+                                    if (isset($styleArray['borders']['vertical'])) {
+                                        $regionStyles['borders'][$innerEdge] = $styleArray['borders']['vertical'];
                                     } else {
                                         unset($regionStyles['borders'][$innerEdge]);
                                     }
@@ -375,7 +373,7 @@ class Style extends Supervisor
             }
 
             // First loop through columns, rows, or cells to find out which styles are affected by this operation
-            $oldXfIndexes = $this->getOldXfIndexes($selectionType, $rangeStartIndexes, $rangeEndIndexes, $columnStart, $columnEnd, $pStyles);
+            $oldXfIndexes = $this->getOldXfIndexes($selectionType, $rangeStartIndexes, $rangeEndIndexes, $columnStart, $columnEnd, $styleArray);
 
             // clone each of the affected styles, apply the style array, and add the new styles to the workbook
             $workbook = $this->getActiveSheet()->getParent();
@@ -387,7 +385,7 @@ class Style extends Supervisor
                 if (self::$cachedStyles === null) {
                     // Clone the old style and apply style-array
                     $newStyle = clone $style;
-                    $newStyle->applyFromArray($pStyles);
+                    $newStyle->applyFromArray($styleArray);
 
                     // Look for existing style we can use instead (reduce memory usage)
                     $existingStyle = $workbook->getCellXfByHashCode($newStyle->getHashCode());
@@ -409,7 +407,7 @@ class Style extends Supervisor
                     if (!$existingStyle) {
                         // The old style combined with the new style array is not cached, so we create it now
                         $newStyle = clone $style;
-                        $newStyle->applyFromArray($pStyles);
+                        $newStyle->applyFromArray($styleArray);
 
                         // Look for similar style in workbook to reduce memory usage
                         $existingStyle = $workbook->getCellXfByHashCode($newStyle->getHashCode());
@@ -428,7 +426,7 @@ class Style extends Supervisor
                         // $newStyle should always be defined.
                         // This block might not be needed in the future
                         $newStyle = clone $style;
-                        $newStyle->applyFromArray($pStyles);
+                        $newStyle->applyFromArray($styleArray);
                     }
 
                     // we don't have such a cell Xf, need to add
@@ -475,33 +473,33 @@ class Style extends Supervisor
             }
         } else {
             // not a supervisor, just apply the style array directly on style object
-            if (isset($pStyles['fill'])) {
-                $this->getFill()->applyFromArray($pStyles['fill']);
+            if (isset($styleArray['fill'])) {
+                $this->getFill()->applyFromArray($styleArray['fill']);
             }
-            if (isset($pStyles['font'])) {
-                $this->getFont()->applyFromArray($pStyles['font']);
+            if (isset($styleArray['font'])) {
+                $this->getFont()->applyFromArray($styleArray['font']);
             }
-            if (isset($pStyles['borders'])) {
-                $this->getBorders()->applyFromArray($pStyles['borders']);
+            if (isset($styleArray['borders'])) {
+                $this->getBorders()->applyFromArray($styleArray['borders']);
             }
-            if (isset($pStyles['alignment'])) {
-                $this->getAlignment()->applyFromArray($pStyles['alignment']);
+            if (isset($styleArray['alignment'])) {
+                $this->getAlignment()->applyFromArray($styleArray['alignment']);
             }
-            if (isset($pStyles['numberFormat'])) {
-                $this->getNumberFormat()->applyFromArray($pStyles['numberFormat']);
+            if (isset($styleArray['numberFormat'])) {
+                $this->getNumberFormat()->applyFromArray($styleArray['numberFormat']);
             }
-            if (isset($pStyles['protection'])) {
-                $this->getProtection()->applyFromArray($pStyles['protection']);
+            if (isset($styleArray['protection'])) {
+                $this->getProtection()->applyFromArray($styleArray['protection']);
             }
-            if (isset($pStyles['quotePrefix'])) {
-                $this->quotePrefix = $pStyles['quotePrefix'];
+            if (isset($styleArray['quotePrefix'])) {
+                $this->quotePrefix = $styleArray['quotePrefix'];
             }
         }
 
         return $this;
     }
 
-    private function getOldXfIndexes(string $selectionType, array $rangeStart, array $rangeEnd, string $columnStart, string $columnEnd, array $pStyles): array
+    private function getOldXfIndexes(string $selectionType, array $rangeStart, array $rangeEnd, string $columnStart, string $columnEnd, array $styleArray): array
     {
         $oldXfIndexes = [];
         switch ($selectionType) {
@@ -514,7 +512,7 @@ class Style extends Supervisor
                     $cellIterator->setIterateOnlyExistingCells(true);
                     foreach ($cellIterator as $columnCell) {
                         if ($columnCell !== null) {
-                            $columnCell->getStyle()->applyFromArray($pStyles);
+                            $columnCell->getStyle()->applyFromArray($styleArray);
                         }
                     }
                 }
@@ -533,7 +531,7 @@ class Style extends Supervisor
                     $cellIterator->setIterateOnlyExistingCells(true);
                     foreach ($cellIterator as $rowCell) {
                         if ($rowCell !== null) {
-                            $rowCell->getStyle()->applyFromArray($pStyles);
+                            $rowCell->getStyle()->applyFromArray($styleArray);
                         }
                     }
                 }
@@ -627,13 +625,13 @@ class Style extends Supervisor
     /**
      * Set Conditional Styles. Only used on supervisor.
      *
-     * @param Conditional[] $pValue Array of conditional styles
+     * @param Conditional[] $conditionalStyleArray Array of conditional styles
      *
      * @return $this
      */
-    public function setConditionalStyles(array $pValue)
+    public function setConditionalStyles(array $conditionalStyleArray)
     {
-        $this->getActiveSheet()->setConditionalStyles($this->getSelectedCells(), $pValue);
+        $this->getActiveSheet()->setConditionalStyles($this->getSelectedCells(), $conditionalStyleArray);
 
         return $this;
     }
@@ -665,20 +663,20 @@ class Style extends Supervisor
     /**
      * Set quote prefix.
      *
-     * @param bool $pValue
+     * @param bool $quotePrefix
      *
      * @return $this
      */
-    public function setQuotePrefix($pValue)
+    public function setQuotePrefix($quotePrefix)
     {
-        if ($pValue == '') {
-            $pValue = false;
+        if ($quotePrefix == '') {
+            $quotePrefix = false;
         }
         if ($this->isSupervisor) {
-            $styleArray = ['quotePrefix' => $pValue];
+            $styleArray = ['quotePrefix' => $quotePrefix];
             $this->getActiveSheet()->getStyle($this->getSelectedCells())->applyFromArray($styleArray);
         } else {
-            $this->quotePrefix = (bool) $pValue;
+            $this->quotePrefix = (bool) $quotePrefix;
         }
 
         return $this;
@@ -716,11 +714,11 @@ class Style extends Supervisor
     /**
      * Set own index in style collection.
      *
-     * @param int $pValue
+     * @param int $index
      */
-    public function setIndex($pValue): void
+    public function setIndex($index): void
     {
-        $this->index = $pValue;
+        $this->index = $index;
     }
 
     protected function exportArray1(): array